requirements.txt
jsonschema
anthropic
python-dotenv